Remove dependency on scipy, numexpr
There were some hidden dependencies on scipy.ndimage and numexpr which
were not recorded in the documentation. This update removes those
dependencies, and replaces the calculation of the slope and aspect with
only numpy functions, based on numpy version 1.6 (which is the standard
packaged with ArcGIS 10.2).

Because of the way that the scipy.ndimage functions calculated the slope
(giving a slightly smoothed result), it was necessary to update the test
data sets. We chose this route to avoid the requirement of essentially
duplicating the scipy.ndimage.convolve2d code, using only Numpy.

Modifications were also made to the way the slope is calculated for cells
adjacent to nodata values. The updated approach gives a better
representation of the slope in these areas that is of a similar scale to
the neighbouring values.